Some important changes to the underlying shinyinputtables framework to make it more workable and maintainable in the future.
1. Moved some of the more general JS functionality into a helper script
and added a data contract Map object to help facilitate JSON data
packaging of event info. The current config seems to bug out RStudio's
build in browser when launching an app, but works as intended when run
externally to browser... need to look into this further.

2. Switched to JQuery event delegation framework for input event
detection. Helps simplify the event manaagement code and results in a
cleaner DOM as we don't need to have an ID for each input cell. Of more
consequence is the fact that binding the event handler to the table, not
individual elements makes the table event mamagement more ammenable to
changes in child element composition (add/remove). Crucially, the Shiny
JS input handling system shouldn't theoretically be "watching" events on
each input cell, only on the main table.

3. The switch in 2. meant that I needed to rework the R code for table
generation, drastically simplifying it and adapting it to be compatible
with the data needs of the new event delegation framework.

There is much more to do, but this set of fixes should help the
performance of shinyinputtables in the near term and make the package
more flexible to the inclusion of anticipated future features.